Cuttoff incidence datatable from interaction data frame In a incidence data.frame, cutoff keeps the maximal incidence information with the minimal number of species possible

Cuttoff incidence datatable from interaction data frame In a incidence data.frame, cutoff keeps the maximal incidence information with the minimal number of species possible

Usage

cutoff_incidence(data, incidence = incidence, accuracy = 4)

Arguments

data

A data.frame object with acummulated interaction information

incidence

The incidence of interaction in the data data.frame

accuracy

The accuracy of the cutoff. Since the cutoff is get from second derivate, the accuracy set the round number of digits to cut the second derivate. Default 5

Value

A data.frame with the summarized incidence from an interaction data.frame. Includes the acummulated incidence for each group provided

Examples

library(ecointeraction)
accumulate_incidence(mammalvirus, virus, incidence) %>% cutoff_incidence
#> # A tibble: 13 × 4
#>       id virus incidence cummulativesum
#>    <int> <chr>     <dbl>          <dbl>
#>  1     1 WNV        1712           1712
#>  2     2 SLEV        759           2471
#>  3     3 JEV         448           2919
#>  4     4 DENV        413           3332
#>  5     5 YFV         319           3651
#>  6     6 WSLV        257           3908
#>  7     7 ZIKV        240           4148
#>  8     8 POTV        178           4326
#>  9     9 ILHV        145           4471
#> 10    10 NTAV         77           4548
#> 11    11 SPOV         71           4619
#> 12    12 BSQV         21           4640
#> 13    13 USUV         18           4658